==> 1.    SCRAPPING_source.R <==

#' @title source scrapping 
#' @description source the script performing the
#' preparatory work on the urls of the documents to
#' download
#' @return a complete dataset with urls of relevant 
#' documents to download for the textmining analysis
#' @author Manuel Betin

source("1.1. SCRAPPING_update_recent_IMF_docs.R")

source("1.2. SCRAPPING_consolidate_urls_IMF_docs.R")
