==> 2.   TEXTMINING_main.R <==
#' @title perform the download of the 
#' pdf reports and the textmining of the 
#' documents
#' @description use the package TextMiningCrisis
#' to perform the NLP analysis on the corpus 
#' of IMF reports
#' @author Manuel Betin, Umberto Collodel
#' @return a database containing the tf_idf for
#' the selected indicators and the underlying 
#' pdf and aggregated corpus

source("2.1. TEXTMINING_exploratory_work.R")

source("2.2. TEXTMINING_run_text_mining.R")
