==> 6. Crisis_network  .Rmd <==
---
title: <center> The PIC index - Measuring Systemic Crisis</center>
runtime: shiny
output: 
  html_document:
    theme: sandstone
    highlight: tango
---

<!--abstract: "This network shows the ocurrrence, the intensity and correlations between macroeconomic shocks as reported in IMF consultations and reports. Use the slider to select the relevant periods and explore the evolution of systemic crisis across years."
 -->
<style>
.main-container {
    max-width: 2000px;
    max-height: 1500px;
    margin-left: 10;
    margin-right: auto;
}
body {
background-color: `r grey.colors(5)[4]`;
}
</style>

A macroeconomic crisis is a complex system it is not possible to derive the collective behavior of the system from the knowledge of its components. The Probability-Intensity-Complexity (PIC) index is a function of 5 components:

  - **Priority:** shocks composing the system
  
  - **Precision:** words defining each shock 
  
  - **Probability:** probability of shocks 
  
  - **Intensity:** severity of shocks
  
  - **Complexity:** inter-connection of shocks 

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = FALSE)
knitr::opts_chunk$set(error = FALSE)
knitr::opts_chunk$set(warning = FALSE)

library(shiny)
library(dplyr)
library(network)
library(GGally)
library(TextMiningCrisis)
library(SetUpProject)
library(PICindex)
library(lubridate)
library(stringr)
library(tidyverse)
library(visNetwork)


background_color=grey.colors(5)[4]
mydata=data("PICdata")
#mydata=rio::import("../Betin_Collodel/2. Text mining IMF_data/datasets/tagged docs/tf_idf_database.RData")
names(mydata)
# observe({
# subsample_countries=SetUpProject::ctry_classif() %>% filter(Classification %in% ifelse(input$mygroup=="All",c("Developped","Emerging"),c(input$mygroup)))
# mydata=mydata %>% filter(iso3c %in% subsample_countries$ISO3_Code)
# })
subsample_countries=SetUpProject::ctry_classif() %>% filter(Classification %in% c("Emerging"))
mydata=mydata #%>% filter(iso3c %in% subsample_countries$ISO3_Code)


```

<br>
```{r PIC index}
renderText({paste0("The PIC index in ",input$myperiods[1]," - ",input$myperiods[2]," is ",
  round(PIC_index(mydata,period_range = c(input$myperiods[1], input$myperiods[2]),
          P_param=1,I_param = 1,C_param=10,min_cor = input$mincorr )$PIC_index*1000,2))
})

```
<br>

<center>
```{r Main figure,fig.align="center"}

renderPlot(network_PIC(mydata,
                            c(input$group1,input$group2,input$group3,input$group4),
                            min_year=input$myperiods[1],max_year = input$myperiods[2],
                            mymode=input$mymode,
                            min_corr=input$mincorr,
                            labelsize=input$labelsize,
                            background_color=background_color
                            ),
           height = 700, width = 1600)
```
</center>
<br>

```{r type visualization}
fluidRow(
  column(4,shiny::selectInput("mymode","Layout of visualization",choice=c("kamadakawai","circle","target")))
  ,column(4,shiny::selectInput("labelsize","Size of label",choice=c("Uniform size","Function of intensity")))
  ,column(4,shiny::selectInput("mincorr","Minimum correlation to display edges",choice=seq(0,1,0.1)))
  #,column(4,shiny::selectInput("mygroup","Group of countries",choice=c("All","Developped","Emerging")))
)
```

<center>
```{r inputs periods}

sliderInput("myperiods",NULL,
            min=1960,
            max=2016,
            value=c(2005,2010),
            step=5,
            width=400,
            dragRange = T,
            animate = animationOptions(interval=4000,loop=T))

```
</center>

<div font-size:0.7em>
**Note:**
*Edge color, type and size measure the correlation between crisis.*
  - *Red links denote correlations higher than 0.5*
  - *Orange links denote correlations between 0.2 and 0.5,*
  - *Dotted green lines denotes correlations between 0.1 and 0.2.*

*The size of node measure the average intensity of each crisis measured as the average of the normalized tf-idf for each of the categories*

*The opacity of the label denotes the probability of occurence measures*
</div>


```{r sample of country, results='asis',eval=F}
paste0((mydata %>% filter(!is.na(Period)))$iso3c %>% unique(),collapse=", ")
```

<hr>

```{r input shocks}
fluidRow(column(3,
                checkboxGroupInput("group1", 
                              h3("Real shocks"), 
                              choices = list("Natural_disaster" = "Natural_disaster", 
                                             "Commodity_crisis" = "Commodity_crisis", 
                                             "Political_crisis" = "Political_crisis",
                                             "Severe_recession" = "Severe_recession", 
                                             "Soft_recession" = "Soft_recession", 
                                             "Inflation_crisis" = "Inflation_crisis", 
                                             "Trade_crisis" = "Trade_crisis",
                                             "Social_crisis"="Social_crisis",
                                             "Epidemics"="Epidemics"
                                             ),
                              selected = c("Natural_disaster", "Commodity_crisis", 
                                           "Political_crisis","Trade_crisis",
                                            "Social_crisis","Inflation_crisis","Severe_recession"))),
         column(3,
                checkboxGroupInput("group2", 
                              h3("Fiscal financial shocks"), 
                              choices = list("Sovereign_default"="Sovereign_default",
                                             "Fiscal_outcomes"="Fiscal_outcomes",
                                             "Fiscal_consolidation"="Fiscal_consolidation"                                             ),
                              selected = c("Sovereign_default"))),
         column(3,
                checkboxGroupInput("group3", 
                              h3("Other financial shocks"), 
                              choices = list( 
                                             "World_outcomes" = "World_outcomes",
                                             "Contagion"="Contagion",
                                             "Expectations"="Expectations",
                                             "Banking_crisis"="Banking_crisis",
                                             "Financial_crisis"="Financial_crisis"
                                             ),
                              selected = c(
                                             "World_outcomes","Contagion", 
                                             "Expectations","Banking_crisis",
                                             "Financial_crisis")
                              
                              )),
         column(3,
                checkboxGroupInput("group4", 
                              h3("External balance shocks"), 
                              choices = list(
                                             "Balance_payment_crisis" = "Balance_payment_crisis",
                                             "Currency_crisis_severe"="Currency_crisis_severe",
                                             "Currency_crisis"="Currency_crisis",
                                             "Reduction_reserves"="Reduction_reserves"
                                             ),
                              selected = c("Balance_payment_crisis",
                                            "Currency_crisis_severe"
                              )))
         )

```
